ml models/resume-parser/requirements.txt
fastapi
uvicorn
python-multipart
pdfplumber
python-docx
spacy